requirements.txt
APScheduler==3.10.1
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
DateTime==5.1
idna==3.4
numpy==1.25.0
pandas==2.0.3
pyodbc==4.0.39
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.3
zope.interface==6.0
